Test against a boto client that doesn't require a region
This was working locally because I have a default region set in my
personal AWS config. But on Travis CI, there are no credentials. Rather
than set a region name for the EC2 client that's superfluous to what I'm
testing, I've opted to use an IAM client instead because it doesn't
require a region name.